requirements.txt
aiofiles>=23.2.1
ansi2html>=1.9.1
anyio>=4.1.0
appdirs>=1.4.4
asgi-csrf>=0.9
asgiref>=3.7.2
asttokens>=2.4.1
attrs>=23.1.0
beautifulsoup4>=4.12.2
blinker>=1.7.0
branca>=0.7.1
bs4>=0.0.1
certifi>=2023.11.17
cffi>=1.16.0
charset-normalizer>=3.3.2
click>=8.1.7
click-default-group>=1.2.4
colorama>=0.4.6
comm>=0.2.0
contourpy>=1.2.0
cryptography>=42.0.0
cycler>=0.12.1
dash>=2.18.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasette==0.64.5
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fanstatic==1.4
fastjsonschema==2.19.0
filelock==3.14.0
Flask>=3.0.3
folium>=0.15.1
fontawesomefree>=6.5.1
fonttools>=4.45.1
frozendict==2.3.9
geographiclib==2.0
geopy==2.4.1
gspread
greenlet>=3.0.2
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore>=1.0.2
httpx>=0.25.2
hupper==1.12
idna==3.6
importlib-metadata==7.0.0
ipykernel==6.27.0
ipython==8.17.2
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.1
itsdangerous==2.1.2
janus==1.0.0
jedi==0.19.1
Jinja2>=3.1.4
joblib>=1.3.2
js==1.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lxml>=4.9.3
MarkupSafe>=2.1.3
matplotlib>=3.8.4
matplotlib-inline>=0.1.6
mergedeep==1.3.4
multitasking==0.0.11
nba_api>=1.5.0
nbformat==5.9.2
nest-asyncio==1.5.8
numpy>=1.26.4,<2.0.0
oauth2client
openpyxl>=3.1.2
packaging>=23.2
pandas>=2.2.0
parso==0.8.3
peewee==3.17.0
Pillow>=10.4.0
Pint==0.22
pipdeptree==2.13.1
pipreqs==0.4.13
platformdirs==4.0.0
plotly>=5.22.0
pluggy>=1.3.0
prettytable>=3.9.0
prompt-toolkit>=3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycoingecko==3.1.0
pycparser==2.21
Pygments==2.17.2
pyodide-py==0.24.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-multipart==0.0.6
pytz==2023.3.post1
pywaffle==1.1.0
PyYAML==6.0.1
pyzmq==25.1.1
randomuser==1.6
referencing==0.31.0
requests>=2.32.0
retrying>=1.3.4
rpds-py>=0.13.1
scikit-learn>=1.4.2
scipy>=1.12.0
seaborn>=0.13.0
setuptools>=69.5.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy>=2.0.30
sqlparse>=0.4.4
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions>=4.12.0
tzdata==2023.3
urllib3>=2.2.0
uvicorn>=0.24.0.post1
virtualenv==20.26.1
wcwidth==0.2.12
webencodings==0.5.1
WebOb==1.8.7
Werkzeug>=3.0.3
wheel==0.42.0
widgetsnbextension==4.0.9
wordcloud==1.9.3
xlrd==2.0.1
xyzservices==2023.10.1
yarg==0.1.9
yfinance==0.2.4
zipp==3.17.0
